Solve compilation issue

+++ include/megaapi_impl.h
@@ -1154,7 +1154,7 @@ class MegaTransferPrivate : public MegaTransfer, public Cacheable
 
         MegaCancelToken* getCancelToken() override;
         bool isRecursive() const { return recursiveOperation.get() != nullptr; }
-        unsigned getTotalRecursiveOperation() const;
+        size_t getTotalRecursiveOperation() const;
 
         CancelToken& accessCancelToken() { return mCancelToken.cancelFlag; }
